change websocket files
